refractor: move Information and Howto component into their own comonent folder
