+++ src/App.js
@@ -13,7 +13,7 @@ import './styles/main.scss'
 import Header from './components/React/Header';
 import Visualizer from './components/React/visualizer/Visualizer';
 import Footer from './components/React/Footer'
-import Information from "./components/React/Information";
+import Information from "./components/React/information/Information";
 
 
 export default function App() {
